+++ src/components/Divider/Divider.tsx
@@ -21,7 +21,7 @@ import { ReactElement } from 'react'
 
 import { Orientation, TextOrientation } from './Divider.types'
 import { DividerProps } from './Divider.props'
-import useTheme from '../../hooks/useTheme'
+import useTheme from '../../hooks/useTheme/useTheme'
 
 const { Horizontal } = Orientation
 const { Left } = TextOrientation
