+++ src/hooks/useTheme/index.ts
@@ -16,23 +16,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { useContext, useMemo } from 'react'
+import useTheme from './useTheme'
 
-import type Theme from '../themes/schema'
-import { ThemeContext } from '../components/ThemeProvider'
-import themes from '../themes'
-
-const { lightTheme: defaultTheme } = themes
-
-/**
- * A hook to access the current theme from the context.
- *
- * @returns {Theme} The current theme.
- */
-const useTheme = (): Theme => {
-  const theme = useContext(ThemeContext)
-
-  return useMemo(() => theme ?? defaultTheme, [theme])
+export default {
+  useTheme,
 }
-
-export default useTheme
