+++ src/hooks/useTheme/useTheme.ts
@@ -16,10 +16,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import useFeedbackMessage from './useFeedbackMessage'
-import useTheme from './useTheme'
+import { useContext, useMemo } from 'react'
 
-export default {
-  useFeedbackMessage,
-  useTheme,
+import type Theme from '../../themes/schema'
+import { ThemeContext } from '../../components/ThemeProvider'
+import themes from '../../themes'
+
+const { lightTheme: defaultTheme } = themes
+
+/**
+ * A hook to access the current theme from the context.
+ *
+ * @returns {Theme} The current theme.
+ */
+const useTheme = (): Theme => {
+  const theme = useContext(ThemeContext)
+
+  return useMemo(() => theme ?? defaultTheme, [theme])
 }
+
+export default useTheme
